move fakes into test submodule
