+++ rust/src/tests/serialization/certificates.rs
@@ -1,8 +1,4 @@
-use crate::fakes::{
-    fake_anchor_data_hash, fake_genesis_delegate_hash, fake_genesis_hash, fake_key_hash,
-    fake_pool_metadata_hash, fake_script_hash, fake_vrf_key_hash,
-};
-use crate::tests::mock_objects::create_anchor;
+use crate::tests::mock_objects::{create_anchor, fake_anchor_data_hash, fake_genesis_delegate_hash, fake_genesis_hash, fake_key_hash, fake_pool_metadata_hash, fake_script_hash, fake_vrf_key_hash};
 use crate::*;
 
 macro_rules! to_from_test {
